+++ internal/client/http.go
@@ -23,26 +23,33 @@ var (
 func init() {
 	// Build known headers and a list of common values.
 	// See more: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers
-	contentTypes := concat("application", "json", "xml")
-	contentTypes = append(contentTypes, concat("text", "html", "csv")...)
+
+	// https://www.iana.org/assignments/media-types/media-types.xhtml
+	application := withPrefix("application", "json", "xml", "gzip", "zip")
+	text := withPrefix("text", "html", "csv", "css", "xml", "")
+
+	contentTypes := make([]string, len(application)+len(text))
+	contentTypes = append(contentTypes, application...)
+	contentTypes = append(contentTypes, text...)
 
 	utc := time.Now().UTC()
 
 	KnownHTTPHeaders = map[string][]string{
-		// https://www.iana.org/assignments/media-types/media-types.xhtml
 		"Content-Type":     contentTypes,
+		"Content-Encoding": {"gzip", "compress", "deflate", "br"},
 		"Accept":           contentTypes,
+		"Accept-Encoding":  {},
 		"Authorization":    {"Bearer ", "Basic "},
 		"X-Correlation-Id": {},
-		"Date":             {utc.Format(time.RFC1123), utc.Format(time.RFC1123Z)},
 		"User-Agent":       {},
+		"Date":             {utc.Format(time.RFC1123), utc.Format(time.RFC1123Z)},
 	}
 }
 
-func concat(name string, variant ...string) []string {
+func withPrefix(prefix string, variant ...string) []string {
 	values := make([]string, len(variant))
 	for _, v := range variant {
-		values = append(values, name+"/"+v)
+		values = append(values, prefix+"/"+v)
 	}
 	return values
 }