+++ internal/server/server.go
@@ -11,6 +11,7 @@ import (
 
 type Server struct {
 	server *http.Server
+	port   uint
 }
 
 func New(port uint) *Server {
@@ -23,6 +24,7 @@ func New(port uint) *Server {
 	}
 	return &Server{
 		server: s,
+		port:   port,
 	}
 }
 
@@ -31,28 +33,30 @@ func (s *Server) Close() error {
 }
 
 func (s *Server) Serve() error {
+	fmt.Printf("Starting server on :%s.\n", style.Bold.Render(fmt.Sprint(s.port)))
+	fmt.Printf("Press %s to exit.\n", style.Bold.Render("CTRL-C"))
 	return s.server.ListenAndServe()
 }
 
 func handler(w http.ResponseWriter, r *http.Request) {
 	fmt.Println("Incoming request:")
-	fmt.Printf("  Method:  %s\n", style.GreenB(r.Method))
-	fmt.Printf("  Path:    %s\n", style.GreenB(r.URL.Path))
+	fmt.Printf("  Method:  %s\n", style.GreenB.Render(r.Method))
+	fmt.Printf("  Path:    %s\n", style.GreenB.Render(r.URL.Path))
 
 	if len(r.Header) > 0 {
 		fmt.Println("  Headers:")
 		for name, values := range r.Header {
-			v := strings.Join(values, style.Grey("; "))
-			v = fmt.Sprintf("%s %s %s", style.GreenB("["), v, style.GreenB("]"))
-			fmt.Printf("    %s: %s\n", style.Bold(name), v)
+			v := strings.Join(values, style.Grey.Render("; "))
+			v = fmt.Sprintf("%s %s %s", style.GreenB.Render("["), v, style.GreenB.Render("]"))
+			fmt.Printf("    %s: %s\n", style.Bold.Render(name), v)
 		}
 	}
 
 	b, err := io.ReadAll(r.Body)
 	if err != nil {
-		fmt.Printf("  Body:    %s: %s\n", style.RedB("error"), err)
+		fmt.Printf("  Body:    %s: %s\n", style.RedB.Render("error"), err)
 	} else if len(b) > 0 {
 		s := fmt.Sprintf("%d bytes", len(b))
-		fmt.Printf("  Body:    %s\n", style.GreenB(s))
+		fmt.Printf("  Body:    %s\n", style.GreenB.Render(s))
 	}
 }