+++ internal/style/style.go
@@ -7,17 +7,17 @@ import (
 var (
 	bold = lipgloss.NewStyle().Bold(true)
 
-	None    = lipgloss.NewStyle().Render
-	Bold    = bold.Render
-	Red     = lipgloss.NewStyle().Foreground(lipgloss.Color("1")).Render
-	RedB    = bold.Copy().Foreground(lipgloss.Color("1")).Render
-	Green   = lipgloss.NewStyle().Foreground(lipgloss.Color("2")).Render
-	GreenB  = bold.Copy().Foreground(lipgloss.Color("2")).Render
-	Yellow  = lipgloss.NewStyle().Foreground(lipgloss.Color("3")).Render
-	YellowB = bold.Copy().Foreground(lipgloss.Color("3")).Render
-	Blue    = lipgloss.NewStyle().Foreground(lipgloss.Color("4")).Render
-	BlueB   = bold.Copy().Foreground(lipgloss.Color("4")).Render
-	Cyan    = lipgloss.NewStyle().Foreground(lipgloss.Color("6")).Render
-	CyanB   = bold.Copy().Foreground(lipgloss.Color("6")).Render
-	Grey    = bold.Copy().Foreground(lipgloss.Color("245")).Render
+	None    = lipgloss.NewStyle()
+	Bold    = bold
+	Red     = lipgloss.NewStyle().Foreground(lipgloss.Color("1"))
+	RedB    = bold.Copy().Foreground(lipgloss.Color("1"))
+	Green   = lipgloss.NewStyle().Foreground(lipgloss.Color("2"))
+	GreenB  = bold.Copy().Foreground(lipgloss.Color("2"))
+	Yellow  = lipgloss.NewStyle().Foreground(lipgloss.Color("3"))
+	YellowB = bold.Copy().Foreground(lipgloss.Color("3"))
+	Blue    = lipgloss.NewStyle().Foreground(lipgloss.Color("4"))
+	BlueB   = bold.Copy().Foreground(lipgloss.Color("4"))
+	Cyan    = lipgloss.NewStyle().Foreground(lipgloss.Color("6"))
+	CyanB   = bold.Copy().Foreground(lipgloss.Color("6"))
+	Grey    = bold.Copy().Foreground(lipgloss.Color("245"))
 )