+++ main.go
@@ -11,12 +11,12 @@ import (
 func main() {
 	cmd, err := cli.Build("v0.12.1")
 	if err != nil {
-		fmt.Println(style.RedB("error:"), err)
+		fmt.Println(style.RedB.Render("error:"), err)
 		os.Exit(1)
 	}
 
 	if err = cmd.Execute(); err != nil {
-		fmt.Println(style.RedB("error:"), err)
+		fmt.Println(style.RedB.Render("error:"), err)
 		os.Exit(1)
 	}
 }