+++ tui/state.go
@@ -2,10 +2,19 @@ package tui
 
 import (
 	"fmt"
+	"io"
 	"net/http"
 	"strings"
 
+	"github.com/charmbracelet/lipgloss"
+	"github.com/lunjon/http/internal/style"
 	"github.com/lunjon/http/internal/types"
+)
+
+var (
+	confirmedPrefix = confirmedStyle.PaddingLeft(1).Render(okIcon + " ")
+	stateFieldStyle = boldStyle.Copy().Width(8).Align(lipgloss.Left)
+	stateValueStyle = style.Cyan.Copy()
 )
 
 type state struct {
@@ -41,34 +50,44 @@ func (s state) render() string {
 	var b strings.Builder
 	if s.method.IsSome() {
 		method := s.method.Value()
-		b.WriteString(fmt.Sprintf("Method:  %s\n", confirmedStyle.Render(method)))
+		renderStateValue(&b, "Method", method)
 	}
 
 	if s.url.IsSome() {
 		url := s.url.Value()
-		b.WriteString(fmt.Sprintf("URL:     %s\n", confirmedStyle.Render(url)))
+		renderStateValue(&b, "URL", url)
 	}
 
 	if s.body.IsSome() {
-		b.WriteString(fmt.Sprintf("Body:    %s\n", s.bodyDescription))
+		renderStateValue(&b, "Body", s.bodyDescription)
 	}
 
 	if s.headers.IsSome() {
 		headers := s.headers.Value()
 		if len(headers) > 0 {
-			b.WriteString("Headers:\n")
-			taber := types.NewTaber("  - ")
+			renderStateValue(&b, "Headers", "")
+
+			taber := types.NewTaber("   - ")
 			for name, values := range headers {
-				key := headerKeyStyle(name + ":")
-				value := headerValueStyle(strings.Join(values, "; "))
+				key := headerKeyStyle.Render(name + ":")
+				value := headerValueStyle.Render(strings.Join(values, "; "))
 				taber.WriteLine(key, value)
 			}
 			b.WriteString(taber.String())
 		} else {
-			b.WriteString("Headers: ")
-			b.WriteString(confirmedStyle.Render("[]"))
+			renderStateValue(&b, "Headers", "[]")
 		}
 	}
 
 	return b.String()
+}
+
+func renderStateValue(w io.Writer, name, value string) {
+	fmt.Fprintf(
+		w,
+		"%s%s %s\n",
+		confirmedPrefix,
+		stateFieldStyle.Render(name+":"),
+		stateValueStyle.Render(value),
+	)
 }